Moves FaspManager under an aspera directory.
